requirements.txt
# Django Framework
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0

# Configuration and Environment
python-decouple==3.8

# AI and Machine Learning
google-generativeai==0.3.2

# HTTP and API clients
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9

# Data validation and serialization
pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0

# Async support
anyio==4.10.0
sniffio==1.3.1

# Utilities
tqdm==4.67.1
distro==1.9.0
jiter==0.10.0

# Type hints and validation
typing-extensions==4.14.1
typing-inspection==0.4.1

# HTTP protocol support
h11==0.16.0
idna==3.10
certifi==2025.8.3
urllib3==2.5.0
charset-normalizer==3.4.3

# Production dependencies
gunicorn==21.2.0
whitenoise==6.8.2
dj-database-url==2.3.0

# PostgreSQL adapter (psycopg3 - better Python 3.13 support)
psycopg[binary]==3.2.3

# Stripe Payment Integration
stripe==8.10.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-django==4.7.0
# black==23.11.0
# flake8==6.1.0

# Optional production dependencies
# redis==5.0.1
# celery==5.3.4
